I wrote 2 new functions: 1. int PAPI_load_event(EventSetInfo *thisESI, int EventSetIndex); 2. int PAPI_load_dataSlotArrayElement(EventSetInfo *ESI);
The second one is a bookkeeping thing that is internally consistent.

The first one -- PAPI_load_event -- I will need help with.  This is
a draft function.  It needs the hardware specific information.
We can talk about this one tomorrow, Phil.

In the past 2 days, I have added and rewritten a lot of stuff.  I wanted
to get this done before Monday so it would be on the table to discuss.
Later today I will go back through and proof read -- try to make the
function names correct in regards to Phil's preferred use of

    papi_  prefix
    _papi_ prefix
    PAPI_  prefix

but for now taking a break.

cheers!
cricket!