Hi all,
This is a major code revision. The P2SC works, but it is not heavily tested.
There is a new argument to the low level add/remove event routines. It
is an index into a new structure called EventMapArray in the ESI. It
holds a map from the index of the user added event to a native piece
of integer information. On the IBM, this index contains if < 7, the
number of the counter corresponding to the added event. If it is > 7
it is the mask of the 7 counters << 3. Don't worry about it for now.
It's aportable attempt at a mask array.

Note, new branch events have been added and we have changed the presets
so the SGI will likely break on these.

Also, I've cleanup up a lot of PAPI code. I've also added a new
directory called p2sc in src which contains kernel/stub routines
necessary for compilation.

Cricket, in order to update the documentation you should look at how
to use cvs diff command. But don't update docs until you finish the tool.

-Phil