More fixes for multithreaded profiling and overflow, plus a PAPI_sprofil() call for doing multiple disjoint text segments
Also, fixed up some test files and replaces handle_error() which
can be called if you compile with -DDEBUG. If you want MEGAverbosity
setenv PAPI_DEBUG in addition to -DDEBUG.

Plus I've begun adding support Linux 2.4 with Mikael Petterson's
perfctr patch. It won't work with threads yet, I've got to
change the perfctr library a bit.

This is not a stable release, so be careful