Addition of PAPI_accum_counters to the High-level API and corresponding documentation updates and example files.
PAPI_accum_counters has exactly the same semantics as PAPI_accum in that
the counters are cleared after the call. This is in contrast to the relation
between PAPI_read_counters and PAPI_read, where the high-level call clears
the counters while the low-level does not. For the easily confused (read
the current contributor) the difference in semantics creates problems.

PAPI_read_counters and PAPI_accum_counters can be used as a simple way
of "holding" the counters over sections of code that you dont want measured.
This is illustrated in the examples provided.

It could be considered a nice thing to do to modify the semantics of
PAPI_read_counters so as it does not clear the counters and thus behaves
more similar to PAPI_read. If this is done, the call PAPI_reset_counters
would probably be a valuable addition to the high-level interface.

/Nils, smeds@pdc.kth.se