This might be a little pre-mature, but in that case I trust that the PAPI group removes the files. This is a simple attempt to adapt the linux-perfctr substrate to the coming 2.0 of the perfctr library.
This initial version uses _no_ new parts of perfctr 2.0. It is more or less
the smallest number of modifications needed to make PAPI at least seem to
work. Hopefully it can be used as a grounds for discussion on what we want
to change for the final 2.0 of perfctr?

Nils Smeds, smeds@pdc.kth.se