Added new code to support the UltraSparc III. None of that code has been tested as we don't have access to the platform. Also, added code for that platform to use the hardware overflow feature of the hardware.
One thing concerns me: Can the USIII count flops?

The preset map for this platform should be considered incomplete as
I don't have all the tech specs yet.

I made a minor chance to the Makefiles.

The first tells the Solaris dynamic linker that libpapi.so requires libcpc.so,
so now you only have to link with the papi library and the runtime linker does
the magic.

The other change was to add the MISCOBJS depend to the shared library line
in Makefile.inc. The cause was that get_tick.o was being left out of the build
for the shared lib.

Enjoy

Phil