Ok, this commit fixes a couple of things.
First and most importantly, is the bug with the use of
PAPI_get_virt calls in a threaded environment on linux.

1) My locking was broken, both in location and in function.

It now works correctly, someone should test it on a
SMP machine tho'. Basically, the locks were never getting
decremented back to zero, so I faced deadlock about 25% of
the time. This was a B*tch to track down, but it now works.

The overhead of the locking, if it needs to wait, is only about
100 cycles, on my 1Ghz PIII here the virtual timer costs about
100 cycles when no lock contention, the spin adds about 100 cycles.

2) Thread initialization/detection routines were not being called
in the virt calls, I fixed that as well.

3) Optimal flags for Linux/GNU builds. Now added
-fomit-frame-pointer giving us another precious register
-funroll-loops for all those dump small loops in the code

4) New multiplex_and_pthreads target that someone started to add, but
the result was it wasn't being built on ANY platform. Shirley did this
commit, my guess was it's unfinished. This is now fixed everywhere.

5) Why can't we build multiplexing on the T3E?

6) Fixed the openmp flags on sun, we need -xopenmp with Sun Forte 6.2 compilers.

Peace love and Tofu,

phil