We've got a dual AMD ;-)
Fixes needed to make PAPI work on this platform applied in this patch.
I only applied changes to code inside #ifdef PERFCTR20 so it will not work
if you are running PERFCTR 1.x, but you were thinking of upgrading anyway
weren't you?

If you are on a Dual AMD you might need to apply manual updates to your
kernel due to what appears as a bug in boot.S.
(Thanks Mikael Pettersson for fixing this for me)

I think that the mapping of native events to PAPI events could possibly
be improved. Who wrote it? Phil?