Added a new function PAPI_add_env_event(int *EventSet, int *EventCode, char *env_variable)
This is much like PAPI_add_event but with one major change,
the env_variable holds the name of the environment variable that holds
the event, you can use preset names like PAPI_FLOPS, the integers which
allows for native events, but 0x80000 format isn't working yet but will
tommorow.  The EventCode is a default Event which will be used if
the env_variable is not set, or is invalid when it tries to add it,
EventCode at return will contain the Event that was added (Either the
enviroment variable, or default EventCode)

This can be useful to change events, during runtime or recompiling with
out that much additional overhead, but because there is some overhead
I called it a different function.