Renewed implementation of "non-decreasing" multiplexing counters. With PAPI compiled with -DMPX_NONDECR (default option) consequtive PAPI_read of the same event will never give decreasing results.
The code in MPX_stop is simplified in this version as MPX_read is used
inside MPX_stop for updating event set internal data structures.

The multiplexing code is prepared for being able to "multiplex" rate
events, but this is untested as there is lacking support in the PAPI
library for running PAPI_FLOPS and PAPI_TOT_CYC simultaneously on the
PII/PIII platform. (This deficiency is fixed in the "register optimization"
scheme under development?)