This is Per Ekman's patch to fix the "events returned out of order" problem that was reported by Frank Deserno. The fix is to rebuild the selector bit pattern after each PAPI_add_event for all events in the event set involved. This is necessary as the underlying libpfm substrate reorders the counters.
For PAPI v3 one should consider to add a pointer in each event sstructure
that points back to the "head" of the event set as this is needed for the
traversal through the events when updating the selectors.

In this fix, a similar functionality is implemented via an auxiliary
function
	EventSetInfo *get_my_EventSetInfo(EventInfo_t *me)
added to papi.c/papi_internal.h

An updated release should be considered (in my opinion) as the current
IA-64 version has the problem of handing counters back in a more or less
arbitrary order. If there is a tag in the CVS corresponding to the current
release we could make a branch and release a bug-fix version, if a full
release from the CVS is too shaky.