Multiplexing bugs fixed in pentium 4 when total cycles were used with total instructions.
The problem? They were sharing the same ESCR register. Total cycles is a special
case so it has been moved to counter 4, which no other event uses.

ddiOptimization has been switched to -O3 and -DDEBUG has been removed. The reason?
This is a production releaqse and will be the last one for a while. Users who experience errors
should recompile with -DDEBUG

-Phil