Ok, from now on, anyone who checks in a #define into a test.c file will be flogged with Udon noodles. All test definitions have been moved to test_utils.h. No more #def NUM in these tests.
The exceptions are the sdsc tests and the hwprofile test (for now.)

In addition, sprofile has been fixed. The AIX/Power architecture uses
function descriptors instead of function pointers, just like the IA64
plaform. If you don't know what that is, check the web. It only happens
when you try to pass an address of a function.

Now, sprofile passes on AIX and overflow passes on Solaris.