I have added multiple locking mechanisms as macros, the new API is:
PAPI_lock(int lock)
PAPI_unlock(int unlock)

The user can use PAPI_USR1_LOCK and PAPI_USR2_LOCK, internally we have
PAPI_INTERNAL_LOCK and PAPI_MULTIPLEX_LOCK

All calls including PAPI_lock are implemented as macros, currently
the max locks we will be able to support is 31 because of Cray T3E,
and while I have tried to implement the calls on all substrates I may
have missed some and I did not try to compile this on the various platforms
so be aware.  The other change is that we should not call PAPI api directly
from substrates and internally, so internally we should use:
_papi_lock(int lock) and _papi_unlock(int lock)

Also, realize that some platforms use _papi_lock_init() to initialize the
locking mechanisms, so the semantics of the calls is undefined before
calling PAPI_library_init which calls the locking initialize.

In addition, we are missing locking mechanisms on Alpha, I am not positive
that I got the Cray T3E right, but I gave it a good first attempt and some
of the assembly locking routines may not be totally correct.