Created functions for platforms that do not support var arg macros (IE Solaris) to use these, define: NO_VARARG_MACRO
Currently we are using the old gcc style of var arg macros, ie:
#define MYMACRO(format, args...) (printf(format, ## args))

The C99 standard is:
#define MYMACRO(format, ...) (printf(format, ##__VA_ARGS__))

Right now the gcc version seems more compatiable, but sometime in the future
we may have to move over.  Also, notice the ## are needed as if args is
empty this will strip the , out. Otherwise the , would stay in.