Edited the function name handling again. There is a better one than __FUNCTION__ which is the GNU standard. C99 defined __func__ as the CPP keyword of choice.
If you are on a platform where the build fails due to this, please
add -DNO_FUNCTION_MACRO to your platform's makefile and check it in.

Added MACROS file which is there to document (for developers) all the MACRO's
we use. We will have to remove all of these when we go to GNU autoconf
anyways, so documenting them now will help us much later on. Please don't
forget to consult this.