Fixed a seg fault due to an overwrite of an old fix - from version 1.220:
Removed the _papi_hwd_ctl() from the PAPI_DEFDOM case in PAPI_set_opt.
This removes a bug that will either result in a seg fault or papi
error when trying to set the default domain.  I'm not sure if this is
the proper fix, but it does remove the bug.  The thinking is that if
you want to change the default domain, then you don't want to change
the current eventset's domain.  If you want to change the current
eventset's domain, you would use PAPI_DOMAIN instead of PAPI_DEFDOM.
So, _papi_hwd_ctl is not needed.