New test case. I found bugs in PAPI/fork() with papiex on apps that fork. This test case I should have written a long time ago. This test does the following:
It tests that when you are using PAPI and you fork(), you need to reinitialize
the library in the new process. Because PAPI has been initialized to talk to the
parent process and it has no idea that the fork occurred. Therefore the data
structures and control information are for another process which is not the
desired behavior in this case.

This test needs to be a no-op for Windows.