Hi,
This is an iterative update to the Makefiles on the linux platform. Joe
correctly pointed out (along with John Mudd of Portland Group) that the
Opteron really wants -fPIC on the shared build line.

Also, I have finally figured out why we needed the 'ln' phase of the
perfctr build and it has been removed. The reason? We were building linux.c
before the Perfctr library, which did the link for us. A simple reordering
of the MISCOBJS in Rules.perfctr did the trick.

Also, why did I not catch this with a clean cvs co? I figured that out as
well. We were using make clean to perfctr, which did not remove the link.
I have changed the clean target (in native_clean) to make distclean, which
properly removes the link and now makes a 'make clobber' functionally identical
to a cold checkout.

These makefile changes may be soon deprecated in favor of new stuff by
Kevin and Haihang.