Now skipped test case according to temporary deprecation of dmem info.
Added the basics for an SSE/SSE2 test case. This will require a bit more
work. Currently it has no papi in it, just the SSE/SSE2 loops. Take a look
at the assembly.

This test case will have to work (and produce correct numbers) for PAPI 3.0
final. We need to find out what is included in FP_INS and what is included
in VEC_INS.

I'm leaning towards only counting Vector Packed floating point INS
in VEC_INS for the 3.0 final release.