Several Check ins as follows: New error code: PAPI_EPERM for cases where you don't have sufficent permissions to execute.  For example running Kernel or Exception count mode on the X1 or SGI machines.  This will also cause test cases to skip instead of failing.
Support for hardware/software mixed profiling.  This doesn't mean you can do
both at the same time, but allows the X1 to use hardware profiling when it can
and software profiling when it can't

The X1 Port:
There are a few technical issues to realize with the X1 port, which I will
further document in a X1 Readme file.
For overflowing/profiling when only overflowing on the p-chip then the X1 port
will use hardware overflowing support, but if overflowing on a e or m chip (alone
or in combination with a p chip) then the port uses software overflow as the m and
e chips don't support hardware overflow.

Overheads are not consistant as it depends on which events you are monitoring, each
chip has to be read by a seperate ioctl call.  This means that if you are monitoring
events from all 3 chips, each start/stop/read requires 3 ioctl calls, and so if you
monitor on only 1 chip at a time you get less overhead.

There is now support for monitoring each m/e chip instead of an aggregated count of all
of them.  PAPI currently doesn't use this, but this should be explored to see if
perhaps we should.

When running in MSP mode, currently we add all the counters from each ssp and return
this value, but because ssp DM applications such as MPI jobs can take advantage of the ssp's
(Running multiple mpi processes on one MSP), it is probably better to come up with a
way to return each individual SSP, Dan and I have talked about possibly extending
PAPI to support multiple dimension values that depend on the substrate it is on, maybe
something to explore.

The Cray X1 does not support shared objects, thus this is not supported.

The presets need to be fleshed out, and decisions on what should be included where,
for example should vector fp ins be included with the scalar fp ins?  In addition,
there are 3 tlb's, vector, instruction and scalar.  How should this map?

Cache information is not supported yet, but will be added at the same time I flesh
out the presets.  In addition, I need to adjust the Makefile to compile a MSP and
SSP version of the library.

Fortran has not been tested yet, as I need to find the fortran compiler on the machine
I have access to.  Everything should work, unless the X1 uses string formats the same
way the T3E does and in this case I would just need to add the proper additional
#defines.

There are some more technical issues, but I will add those to the X1.README file that
I will add later.

Cheers,
Kevin