Major rewrite (first pass) of pepi_event_info functionality. Data structure changes have been implemented in the independent layer and tested on Pentium 4 and POWER3. No other changes should be required in other substrates (for now).
CAVEAT: To conserve memory footprint and string copies, the info in the
hwi_search_t structures on each substrate ARE ASSUMED TO BE CONST.
Therefore, those platforms (Itanium??) that use non-standard event arrays
should be examined to guarantee that the structure passed to
_papi_hwi_setup_all_presets() is NOT MODIFIED after initialization.

New functionality implemented in an interim version of avail includes
the display of derived type, postfix operators, and developer notes.
It also includes hex codes and names of all native events in a preset.

Still to come: modifications to native event info to allow the display of
hwd_register_t contents. This will require more significant changes in
each substrate.