This fixes the problem with not being able to read the counters inside a handler. The only "real" changes are in perfctr-p3.c and perfctr-p4.c.
Before a handler is called now, the EventSet is set the PAPI_PAUSED and then
that flag is removed once the handler returns.  Then a flags is sent into the
_papi_hwd_read() calls that has the state of the EventSet and the read can use this
to determine how to read the counters if it makes a difference (Which it currently
does for perfctr).  If this is changed later, the PAUSED/flags stuff could be removed
if needed.