Major restructuring of the way we handle READMEs and FAQs. Thanks to Phil for taking the lead on this. All content for README and platform specific READMEs have been merged into a single FAQ. That FAQ is maintained EXCLUSIVELY in the PAPI website database. If you need to ADD or EDIT information to the FAQ **DO IT ON THE WEBSITE**
David Rogers has added a handy-dandy button in the FAQ editior that will
generate a stand-alone fully formatted printable FAQ html file.
That's where this file came from.
In this way, we can distribute the FAQ with the tarball. It will be viewable
and printable from any browser. And it'll be in CVS. And on the website.
But it needs to be maintained from the website interface.