Updated lots of Opteron cache metrics and branch metrics. The only remaining questions are PAPI_L2_TCH and PAPI_L2_TCM. Thanks to AMD and Anoop Iyer for the preset mappings. VENDORS TAKE NOTE. IF YOU HAVE NOT HELPED US GET YOUR PRESET MAPS RIGHT, THEY MAY NOT BE CORRECT!
The following metrics have been updated:
PAPI_MEM_SCY: cycles stalled waiting for memory accesses. The counter
used is in PAPI d8h, which counts dispatch stalls caused by the LS unit
being full. These are not the same. The processor can spend cycles
waiting for memory accesses without the LS unit being full. We do not
have a native counter or a combination of counters that can accurately
count cycles stalled waiting for memory accesses.

PAPI_BR_UCN: unconditional branch instructions. The counter used in PAPI
is c6h, which counts retired far control transfers. This is incorrect;
far control transfers can be conditional or unconditional. We do not
have a native counter or a combination of counters that can accurately
count unconditional branch instructions.

PAPI_BR_CN: conditional branch instructions. The counter used in PAPI is
c2h, which counts retired branches including exceptions and interrupts.
Exceptions and interrupts are not conditional branches, they are
unconditional. We do not have a native counter or a combination of
counters that can accurately count conditional branch instructions.

PAPI_BR_NTK: conditional branches not taken. This quantity is derived in
PAPI as the difference of branches (c2h) and taken branches (c4h). The
problem is the same as that above; there is no way to count conditional
branches.

PAPI_BR_PRC: conditional branches correctly predicted. Ditto.

PAPI_BR_INS: branches. This quantity is derived in PAPI as the sum of
counter c2h (retired branches including exceptions and interrupts) and
c6h (retired far control transfers). The correct way is to use c2h alone.

PAPI_FP_STAL: cycles the FP units are stalled. The counter used in PAPI
is 01h, which actually counts cycles when the FP units are idle (as
opposed to stalled). This counter is correctly used in PAPI for
PAPI_FPU_IDL. There is no way to count cycles when the FP units are stalled.


Enhancements

Some cache events in K8 include speculative activity. We recommend
revamping the cache stats as follows:

PAPI_L1_DCA: 40h. No change.

PAPI_L1_DCM: 41h. We recommend calculating this as the sum of 1e42h and
1e43h, which are DC refill from L2 and DC refill from system respectively.

PAPI_L1_DCH: 40h - 41h. We recommend calculating this as 40h - (1e42h +
1e43h); see previous paragraph.

PAPI_L2_DCA: 41h. We recommend calculating this as the sum of 1e42h and
1e43h, which are DC refill from L2 and DC refill from system respectively.

PAPI_L2_DCM: 27eh. We recommend using 1e43h, which counts DC refill from
system (which is the same as DC refills which missed in the L2).

PAPI_L2_DCH: 1f42h. No change.

PAPI_L1_ICA: 80h. No change.

PAPI_L1_ICM: 81h. We recommend calculating this as the sum of 82h and
83h, which are IC refill from L2 and IC refill from system respectively.

PAPI_L1_ICH: Currently blank. We recommend calculating this as 80h -
(82h + 83h); see previous paragraph.

PAPI_L2_ICA: 81h. We recommend calculating this as the sum of 82h and
83h, which are IC refill from L2 and IC refill from system respectively.

PAPI_L2_ICM: 17eh. We recommend using 83h, which counts IC refills from
system (which is the same as IC refills which missed in the L2).

PAPI_L2_ICH: Currently blank. We recommend using 82h, which counts IC
refills from L2.

(Opteron and Athlon-64 processors have a unified L2 cache. When I say
L2_DCH above I am referring to hits in the unified L2 resulting from L1
DC misses. This is in line with the counters already in PAPI.)


Additions

The following stats which are not computed by PAPI can be computed using
combinations of the above counters. Some of them require more than four
native events, so the calculation is not feasible without multiplexing.

     * PAPI_L1_ICH
     * PAPI_L2_ICH
     * PAPI_L2_ICR
     * PAPI_L2_ICW
     * PAPI_L1_TCH
     * PAPI_L2_TCH
     * PAPI_L2_TCA
     * PAPI_L1_TCR
     * PAPI_L2_TCR