This is a fix for Brian Wylie's bug in AIX/Power.
The issue? A STATIC VARIABLE definition in _papi_hwd_read. It was
causing BOGUS results on HEAVILY threaded programs. Thanks to
Brian for catching this one. We should have a test case for this!

Those of you who don't want to download the CVS version should change this
line in aix.c
<    static pm_data_t data;
---
>    pm_data_t data;