Bug fixes from Maynard Johnson of IBM maynardj@us.ibm.com
1. The patch to extras.c ensures that _papi_hwi_using_signal is always
incremented.
2. The patch to multiplex.c prevents the PAPI_EINVAL error if
MPX_cleanup is called multiple times for a given set of MPX events
(which happens in ctests/multiplex1).
3. The patch to papi_internal.c is to ensure that all native events that
we fail to add during a given call to add_native_events() get cleaned
out of the NativeInfoArray.  This bug was causing a particular testcase
to fail for me. In the testcase, I successfully added a couple events,
then failed while adding a PAPI present event that mapped to 4 native
events.  This change makes sure we examine all NativeInfoArray event
entries to see if they failed and need to be removed.