Changed to less restrictive redistribution clause which brings IBM, Cray, Sun, Pacific Sierra, Etnus, myself, Nils Smeds and others in compliance.
Basically, you can do anything you want with PAPI, including making it a
product and making money off it. What you can't do is you can't change PAPI
and redistribute it as the original PAPI, it must be flagged or marked in
a way that you have changed it, AND you must give credit to the original
authors and ICL/UTK.

This is funny. I wonder if I can get fired for violating my own license?
8-)

Ideally, we should move to stock LGPL, but we'll have to tackle that as
a group.