Fix for aix.c and solaris-ultra in update_shlib_info.
fopen() was not being checked for errors, causing coredumps when
fopen was getting EPERM, Permission Denied. Please, please always check
error codes.

update_shlib_info is a case where this substrate function can fail,
but PAPI should continue to operate. So the update_shlib_info function
returns early with PAPI_OK but notes an error with PAPIERROR macro
which gets dumped if VERBOSE is enabled.

A review of these functions is in order, as there is no consistency checking
being done. If it reads zero lines from the shlib map, the function will
also likely dump core as it doesn't check for a number of valid entries.

Argh.