Addition of papi memory calls.  This allows papi to track memory and free up all memory in a shutdown call.  All calls should now be papi_free, papi_malloc, papi_realloc and papi_calloc
By default papi tracks memory and frees things up.  If:
PAPI_DEBUG_MEMORY is defined then in addition it tracks where the
memory was allocated from and if there is a memory leak at shutdown
prints out the file and line it came from and how many bytes it was.
In addition it also checks for overwritting bounds, by padded the end of the memory with a pattern.  All memory is checked for overwrites with every call to memory (papi_malloc, papi_free, etc...) If you want to force a check you can call check_memory_for_overflow() and it will print out if the memory has been overwritten and where it was allocated from.

If PAPI_DEBUG_MEMORY_TRACE is defined in addition to the functionality defined in PAPI_DEBUG_MEMORY each time a memory call is executed it is logged to stderr. (Location/Line/Size/Pointer value)

If PAPI_NO_MEMORY_MANAGEMENT is specified then the papi_malloc, papi_free, papi_realloc and papi_calloc calls all get redefined to the normal malloc, etc... calls.

Also, because I use a negative offset to access memory, if you call papi_realloc or papi_free on memory not allocated with a papi_ call then the library will coredump. (This is done for speed purposes)  If you have some memory that you want to free, but are unsure if the library or a user supplied the memory, call papi_valid_free() which will only free the memory if the library allocated the memory.  This turns into a noop if PAPI__NO_MEMORY_MANAGEMENT is defined.

Last, there is a new function called papi_mem_overhead(int type) where type can be PAPI_MEM_LIB_OVERHEAD, PAPI_MEM_OVERHEAD or both | together.  The PAPI_MEM_LIB_OVERHEAD returns the memory allocated by the library, the PAPI_MEM_OVERHEAD returns the memory allocated for tracking the memory allocations, and so both | together returns the total memory allocated by the PAPI library.