Added two functions: PAPI_get_substrate_opt which is just like PAPI_get_opt, but takes a substrate as an argument to determine which substrate you are talking about.
PAPI_get_number_hwctrs which is just like PAPI_num_hwctrs but takes
a substrate index.

Ideally, I would have just added new arguments to the old functions.
But breaking backward campatibility with an established user base is
bad ;)


Still need to add Fwrappers for them if we want that.