The attached patch fixes a bug reported to me by Jordi Caubet.  He found that, on the Linux/PPC64 platform, any attempts to change the domain on an event set from "user" to "kernel" were not working.  He provided a code patch with his report.  I enhanced the ctests/second testcase to detect the original problem, and then retested with Jordi's patch applied.  It's working correctly now.
P.S.  The attached patch was created against the papi 3.2.1 release.
The patch applies cleanly to current CVS, with the minor exception of
hunk #1 for perfctr-ppc64.c is offset by 7 lines.  The reason I didn't
bother to port the patch over to current cvs is that what I pulled down
today fails to build.  I'll post the details in a separate note.

Thanks.

Regards,
-Maynard Johnson