iMore PPC32 support. Dan still has to vectorize this.
Note, I REALLY HATE linux.c
It is not linux,it has all sorts of arch specific defines
as well as PERFCTR isms and tons of #ifdefs.

The goal of this file was linux specific but not arch specific
or substrate specific functions. Even hwd_dispatch is in here which
uses vperfctr calls. Bad bad bad.

Note to developers, take this file and SHOOT it for PAPI 4.0.