replaced a structure copy by *a = b (!!!!) to memcpy.
There is still a leak in this code on line 52. Not sure where or who is
supposed to clean this up.

Note that also there can be leaks if memory is allocated inside of
init_substrate.

Ideally, we could probably keep ref counts or else use a global shutdown routine.