Fixes for IRIX. sdsc is still bailing, but subinfo is reporting the correct data. zero passes as well.
IRIX has kernel multiplexing and something seems to be
interfering with that. IRIX should never use PAPI soft
multiplexing, but it seems the conditional code to do this
has disappeared somewhere. Multiplexing on this platform
still needs investigation.

In actuality, this platform should report 2 counters. But as with
before, it reports the number of counters that the kernel multiplexes
namely 32.

What should happen is that the higher level code should check the
kernel multiplex bit and do things as appropriate there. num_counters
should be set to the actual number of counters. We should probably have
a field in sub_info that says the maximum number of kernel multiplex counters.