Fixes for seriously broken Altix kernel substrates. Also support for reset during overflow on IA64 systems. Working and tested on 2.6.16 Altix, 2.6.9 and 2.4.21. Tested on Madison hardware.
There are multiple bugs on the Altix.
1) PFM_END_MSG is sent even when self monitoring for multithreaded exes.
2) PFM_END_MSG can arrive before a PFM_OVFL_MSG for the same thread for
multithreaded exes.
3) Threads occasionally fail to honor a call to PFM_RESTART. Please report
this case should you see this in practice. ONe out of every dozen runs
of ctests/overflow_pthreads will show this error. No workaround for this.
The result of this is that the monitored thread stops getting interrupts
even though overflow is set.