Redefine the Cray X2 counter events structure. Mostly deals with the way M chip events are represented. Previously the libpfm unit masks were used to idenfify which memory chip the event was activated upon. Using unit masks in this way was knowingly incorrect, and tools such as PAPI assume that unit masks can be bit-wise OR'd.
Remove the use of the unit masks for the M chips, and explicitely add the
chip number via "@<n>" where <n> is the chip number 0..15. This is appended
to the base name of the M chip event to fully qualify it.

When events on the M chip are selected, zero unit masks will be enforced.
As before, P and C chip also do not use unit masks.