Ok, this fixes the --with-perfmon argument to be consistent.
The argument is now specified as a true version number, i.e. 2.3, 2.4 and is checked
for format.

If this is not specified and we are not cross compiling, the script looks at
/sys/kernel/perfmon/version and /proc/perfmon to find the current version.

If anything is found over 2.0, it sets the PFMLIB_VERSION_xx where xx are the
two digits.

Tested on:
 IA64/kernel 2.4.x (perfmon 1.6)
 IA64/kernel 2.6.x (perfmon 2.0)
 PIV/kernel 2.6.25+perfmon2 patches (perfmon v2.8)

Attention Cray folks, you'll need to pass --with-perfmon=2.3 or 2.2