The length of the device in the maps file can be longer than just the five bytes that are used by the sscanf() to fetch from the character buffer. Increase the size of the array to hold the device name to 16 and read it from the buffer as a "%s". Also initialize the integral values to zero before sscanf() reads them out of the buffer to better identify invalid acquistion of data.
This changes now ensures that the maps files for Cray XT and X2 compute
nodes are read correctly and test/data/bss address/size details are correct.

This was applied to just the perfmon-version of reading the maps file, but
it probably can be applied to other platforms that also read the file.

Also fix a GNU complaint about paranthesis.