This file addresses the packaging issue reported by SUSE and RedHat where PAPI always builds both shared and static libraries.
Now, the options to configure all do the proper thing.

--disable-shared
--disable-static
--with-shlib=yes/no

The latter says which tests should link with. Of course, it also does checking
in case you request one which it cannot provide.

In addition, some other fixes:

-D_REENTRANT added to Linux builds. THIS IS IMPORTANT.
- install-man uncommented to allow people to install outdated man pages
- The test operator [[ ]] deprecated to use the test command. Remember folks,
  some people may not be running bash, which means you get errors when you run
  this. If you use the test command, you are always safe.

This patch should apply cleanly to PAPI HEAD, which I will submit to the papi
group independently.