Fix compilation error on arches with no user space pmu register
Allow perf_examples/self_count.c to compile on architectures where user
level reads to the performance counters are either not implemented
yet, or is not possible on that architecture.  It appears that it
was intended that the code compile under these circumstances, but
the macro "barrier()" was not defined in these cases.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>