Allow specifying multiple groups to perf_examples/task
- Allow specifying multiple groups to perf_examples/task via multiple
  -e <evt>[,<evt>]... switches
- Remove -g option because it's now the default.  Every -e switch
  gets its own group.
- Add -f option to specify using the experimental PERF_FORMAT_GROUP
  option.  PERF_FORMAT_GROUP appears to be broken in the 2.6.32
  and 2.6.33 when monitoring processes other than the current
  process.  I added this option instead of using a #ifdefs so that
  the code stays live and to avoid maintenance issues down the road.

Because the default behavior is to allow multiple groups, there is no
longer a need for the "read_single()" function, so I have eliminated
it.

This change simplifies the group leader handling and unifies the
logic so that grouping is always used... only the size and
number of groups varies.

To achieve these goals, the profiles to perf_setup_argv_events
and perf_setup_list_events needed to be changed to add num_fds
"in/out" parameters, so that the code can be reentrant.

This change rippled through the other example programs, and to
make the code more clear, I changed the usage of "num" to "num_fds",
"num_events" to "num_fds".  In some of the examples, these needed to
be thread-local storage, and in other cases (syst.c) there is an array
of them, indexed by cpu number.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>