Fix error with buffer size and number of events in a group
perf_get_group_nevents was returning the wrong number of events
when the group was the last in the list - it always returned 1
instead of the true number.

Also, when reading up event groups, only a "max size" (sz) was
being passed to the read() function, which resulted in getting
fewer bytes than requested and a resulting warning when groups with
fewer events follow groups with more events.

Signed-off-by: Corey Ashford <cjashfor@linux.vnet.ibm.com>