Add example component.
I named it "example" as null seemed to not really apply, and
"sample" sounds like it has to do something with counter sampling
which would have been confusing.

This component has three counters, one that returns always 0,
one that returns a constant, one that returns an auto-incrementing value.

The code is commented the best I could (DOxygen when applicable) and has been
run through indent.

While it works as is, there are still some improvements that can be made:
	* PAPI_write() should be implemented
	* auto-increment counter should only increment when it itself is being
		read, and not increment whenever any of the example counters are read.
Both of the above will make the component more comlpicated, but might help
writers of new components.