fix pfmlib_parse_event() to correctly handle event with equivalence
There was a bug introduced by a recent commit which caused events
which equivalence (e.g., UNHALTED_CORE_CYCLES) to fail to encode.
This was due to the fact that the equivalent event was searched for
in all PMU event tables and not just the current one. Some events
may exists in other tables but not necessarily with the same umasks
or encodings.

Signed-off-by: Stephane Eranian <eranian@gmail.com>