make pfm_get_event_first()/pfm_get_event_next() return all supported events
This allows us to encode/validate list events beyond those supported
by the host hardware. The logic is that if the event string contains
a specific pmu type, e.g., wsm::inst_Retired:any_p, then the
library only looks in that PMU event table, otherwise, it looks
in the activated event tables only. An event table is activated when it is
supported by the host hardware or software environment.

The patch updates showevtinfo to allow listing, encoding events with
pmu specifiers.

The man pages for the two entry points are updated to reflect the
change.

Signed-off-by: Stephane Eranian <eranian@gmail.com>